requirements.txt
lime==0.1.1.32
Flask==1.0.2
pandas==0.23.4
matplotlib==3.0.1
numpy==1.15.3
ipykernel==5.1.0
requests==2.18.4
Pillow==5.3.0
beautifulsoup4==4.6.3
scikit_learn==0.20.0
